02-Mistral-Small/requirements.txt
streamlit==1.32.0
mistralai==0.0.12
Pillow==10.2.0
